+++ create.sh
@@ -6,6 +6,6 @@ docker rmi -f petrjahoda/sklabel_cutting_webservice:latest
 docker build -t petrjahoda/sklabel_cutting_webservice:latest .
 docker push petrjahoda/sklabel_cutting_webservice:latest
 
-docker rmi -f petrjahoda/sklabel_cutting_webservice:2020.3.3
-docker build -t petrjahoda/sklabel_cutting_webservice:2020.3.3 .
-docker push petrjahoda/sklabel_cutting_webservice:2020.3.3
+docker rmi -f petrjahoda/sklabel_cutting_webservice:2020.4.1
+docker build -t petrjahoda/sklabel_cutting_webservice:2020.4.1 .
+docker push petrjahoda/sklabel_cutting_webservice:2020.4.1