+++ js/origin.js
@@ -1,0 +1,1 @@
+setTimeout(() => window.history.back(), 30000)