+++ order_scanning.go
@@ -7,6 +7,8 @@ import (
 
 func orderScanning(writer http.ResponseWriter, request *http.Request, params httprouter.Params) {
 	logInfo("Order Scanning", "Page loading...")
-	http.ServeFile(writer, request, "./html/order_scanning.html")
+	writer.Header().Set("Cache-Control", "no-cache, no-store, must-revalidate")
+	writer.Header().Set("Pragma", "no-cache")
+	writer.Header().Set("Expires", "0")
 	logInfo("Order Scanning", "Page loaded")
 }